Added adaptive trapezoidal integration

+++ utility/include/mrock/utility/Numerics/Integration/TrapezoidalRule.hpp
@@ -3,7 +3,7 @@
 
 namespace mrock::utility::Numerics::Integration {
 	template <class UnaryFunction, class RealType, class... Args>
-	auto trapezoidal_rule(const UnaryFunction& function, const RealType begin, const RealType end, unsigned long num_steps, Args&&... args) {
+	auto trapezoidal_rule(const UnaryFunction& function, const RealType begin, const RealType end, unsigned int num_steps, Args&&... args) {
 		const RealType step = (end - begin) / num_steps;
 		decltype(function(begin)) value = 0.5 * (function(begin, std::forward<Args>(args)...) + function(end, std::forward<Args>(args)...));
 		for (int n = 1U; n < num_steps; ++n) {
@@ -14,10 +14,10 @@ namespace mrock::utility::Numerics::Integration {
 	}
 
 	template <class UnaryFunction, class RealType>
-	auto trapezoidal_rule(const UnaryFunction& function, const RealType begin, const RealType end, unsigned long num_steps) {
+	auto trapezoidal_rule(const UnaryFunction& function, const RealType begin, const RealType end, unsigned int num_steps) {
 		const RealType step = (end - begin) / num_steps;
 		decltype(function(begin)) value{ 0.5 * (function(begin) + function(end)) };
-		for (unsigned long n = 1U; n < num_steps; ++n) {
+		for (unsigned int n = 1U; n < num_steps; ++n) {
 			value += function(begin + n * step);
 		}
 
@@ -26,14 +26,14 @@ namespace mrock::utility::Numerics::Integration {
 	}
 
 	template <class UnaryFunction, class RealType>
-	auto trapezoidal_rule_kahan(const UnaryFunction& function, const RealType begin, const RealType end, unsigned long num_steps) {
+	auto trapezoidal_rule_kahan(const UnaryFunction& function, const RealType begin, const RealType end, unsigned int num_steps) {
 		const RealType step = (end - begin) / num_steps;
 		decltype(function(begin)) y{ 0.5 * (function(begin) + function(end)) };
 		decltype(function(begin)) t{ y };
 		decltype(function(begin)) c{ t - y };
 
 		decltype(function(begin)) value{ t };
-		for (unsigned long n = 1U; n < num_steps; ++n) {
+		for (unsigned int n = 1U; n < num_steps; ++n) {
 			y = function(begin + n * step) - c;
 			t = value + y;
 			c = (t - value) - y;
@@ -47,7 +47,7 @@ namespace mrock::utility::Numerics::Integration {
 	template <class Vector, class RealType>
 	auto trapezoidal_rule(const Vector& fx, const RealType step) {
 		auto value = 0.5 * (fx.front() + fx.back());
-		for (unsigned long n = 1U; n + 1U < fx.size(); ++n) {
+		for (unsigned int n = 1U; n + 1U < fx.size(); ++n) {
 			value += fx[n];
 		}
 		value *= step;
